Bump Puma version
We are investigating memory leaks and when looking at the running processes on
the box, we can see that Puma is using more memory than anything else.

There is an issue about this on Github:
https://github.com/puma/puma/issues/1600

This may not be the cause but worth seeing if this stops the memory leaks.