Ensure PDF answers file is the first attachment
Before grouping attachments up to the maximum allowed per email, the pdf
answers file was always the first email to be sent.

It is not quite the same thing, but here we make sure that this answers
pdf is the first attachment in the email. This is on the assumption that
AWS SES keeps the same order of attachments that it receives.