Adjust maximum email size limit for submissions
The [AWS SES docs](https://aws.amazon.com/ses/faqs/#Limits_and_Restrictions)
state that the maximum limit of 10MB also includes the contents of the
email, not just the attachments.

We are leaving 1MB headroom for email contents, which is very generous
as we do not currently use any images etc.