Do not retry email submissions that succeeded previously
Previously when an email failed to send in a Delayed::Job it would retry ALL
emails including, those which previously succeeded, up to the maximum current
retry limit.

Now we write each emails payloads to the DB and whether or not the email
was sent successfully. Unsuccessful emails will continue to be retried
but those which previously succeeded will be ignored.