Differentiate between submission ID and payload submission ID
As a result of keeping track of submission email payloads we wanted to
record the ID of the submission in the database with the EmailPayload.

The submission_id that the JSON and the Email Output Service use for end
users is the submission_id from the payload received from the runnner.
This is not the same as the ID of the submission record in the database
that is created for that given submission.

Here we rename the submission_id received from the payload to
payload_submission_id. We really want to use the term submission_id when
it relates to the ID of a Submission in the database. This will help us
differentiate between the two types.