Enable receipt of CSV submissions from the V2 Runner
This adds the ability for the submitter to handle CSV actions in
submissions received from V2 Runners.

We add a submission_at property to the expected schema as well since
this is used in the CSV file output.

The structure of the payload differs significantly from the legacy
Runner as the V2 Runner does not convert the data and add it all the
actions payload for the CSV. The expectation is that the Submitter will
do the conversion therefore removing the need for the users answers to
appear more than once in the payload, like the do in the legacy Runner.