Use MoJ Forms default no reply email as from address
This is because we have been getting reports of undelivered messages
when sent to certain email clients when using certain From Addresses
that have been set by the user.

Until a more long term fix is found use the MoJ Forms default no reply
address and set the Editor user's From Address as the Reply to address.

If the from address is the same as the MoJ Forms default no reply
address then do not set a reply to address in the payload sent to AWS.