Reply-To: Use the correct values for from and reply-to
The Runner will now pass the Submitter the service name with the
email address, ie: "Service Name <service.name@example.com>".

For a service email, the from address will be the default no-reply
MoJ email address.
For a confirmation email, the from will be a supplied email address.
Therefore, all we need to do in the submitter is accept whatever is
provided in the `opts[:from]` key.

We also only want to use the reply-to field if the `opts[:from]`
provided from the Runner is not the default moj email address.