Ensure only the last full stop is used for filenames
A bug has been reported where if the attachment filename has more
than one occurence of a full stop, it appears as the wrong name
in the final email attachment. Eg:
Filename: 'Screenshot 2023-04-04 at 16.02.20.png' would appear as
'Screenshot 2023-04-04 at 160220.png' in the final email.

This commit fixes this bug.