chore: Tidy up and test coverage increase (#889)
* DRY docker-compose.yml variables

And use same postgres version as the RDS one, as this had diverge too much.

* Increase health controller test coverage

* Increase mock_amazon_ses_adapter test coverage

* Increase attachment_parser_service test coverage

Additionally fixed what I think was a bug, where if some of the attachments failed to parse (missing keys) we log an error but the return result of this logging was being added to the collection (an integer), which I presume is not what we want here.

I've changed the code slightly so the rescue returns `nil`, and using `filter_map` to take care of these nil values automatically.

I've added a test for this edge case scenario too.

* Increase webhook_attachment_service test coverage

In reality this was done by way of removing an unnecessary rescue branch, as the attachments are instances of Attachment (value-object) and they always respond to the methods used here.

* Increase pdf_payload_translator test coverage

In reality this was done by way of removing an unused private method.

* Increase jwt_authentication test coverage

Another unused method it seems. I couldn't find any place where this private method is being used/called.

* Fix deprecation

* Increase error_handling test coverage

And add missing locales (these locales have been copied and are the same as in other repos where they are used).

* Increase process_submission_job test coverage

And configure `simplecov` with a 100% minimum coverage going forward. We can revisit this if we feel is too stringent but I think is a good idea.